% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Data Preparation Function}
\usage{
prep_data(latest_vaccs)
}
\arguments{
\item{latest_vaccs}{A csv file giving the latest data on vaccinations with the following collumns required: iso_a2, iso_a3, Longitude, Latitude, Vaccinated and one collumn per vaccine manufacturer}
}
\description{
A function to merge the latest vaccination data with country polygons
}
\keyword{datawrangling}
