% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{render_bar_plot}
\alias{render_bar_plot}
\title{render_bar_plot}
\usage{
render_bar_plot(sel_lang, relative, vacc_data)
}
\arguments{
\item{sel_lang}{language for map labels. Currently supporting English "en" and German "de".}

\item{relative}{Boolean value to control whether the function outputs a barchart with relative or absolute numbers}

\item{vacc_data}{dataset returned from \link[polivax]{prep_data} function}
}
\value{
shiny.render.function to return plotly barchart
}
\description{
This function plots an interactive bar chart using plotly to show number or proportion of people vaccinated within a country
}
