% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{render_vacc_map}
\alias{render_vacc_map}
\title{render_vacc_map}
\usage{
render_vacc_map(sel_lang, latest_vaccs)
}
\arguments{
\item{sel_lang}{language for map labels. Currently supporting English "en" and German "de".}

\item{latest_vaccs}{A csv file giving the latest data on vaccinations with the following collumns required: iso_a2, iso_a3, Longitude, Latitude, Vaccinated and one collumn per vaccine manufacturer}
}
\value{
shiny.render.function to return leaflet map
}
\description{
This function plots an interactive leaflet map with color polygons of vaccine approvals
}
